% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgnews-package.R
\docType{package}
\name{pkgnews}
\alias{pkgnews}
\alias{pkgnews-package}
\title{pkgnews: Retrieve R Package News Files}
\description{
Read R package news files, regardless of whether or not the
package is installed.
}
\author{
\strong{Maintainer}: Owen Jones \email{owenjonesuob@gmail.com}

}
\keyword{internal}
