% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgnews.R
\name{show_pkg_news}
\alias{show_pkg_news}
\title{Access package news files}
\usage{
show_pkg_news(pkg, gh_release = FALSE)
}
\arguments{
\item{pkg}{the name of the package}

\item{gh_release}{flag whether or not to source the NEWS file from the local R
library or whether to try to direct you to the Github releases page if the
Github URL is available in the package's \code{DESCRIPTION} file (default:
\code{FALSE}).}
}
\value{
a call to \code{[utils::browseURL()]} to open the NEWS file or an
appropriate error message if this fails
}
\description{
R package NEWS files are named \code{NEWS} but can have three particular formats:
plain text files with no file extension; plain text markdown files with the
extension \code{.md} (i.e. \code{NEWS.md}); or R documentation files with the extension
\code{.Rd}. The last format is very rare. While it is reasonably simple to access
the first two types of file (e.g. call to \code{news(package = "curl")}), getting
access to the \code{NEWS.md} file is a bit more verbose. This function makes
this more parsimonious and allows you to access the package's Github release
pages where none of these are available.
}
\examples{
\dontrun{
show_pkg_news("curl")  # NEWS
show_pkg_news("Rcpp")  # NEWS.Rd
show_pkg_news("readr") # NEWS.md
}
}
